+++ src/main/java/pl/danielstawiarski/githubtask/service/CalculationService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 public class CalculationService {
 
     //TODO There is no requirement regarding 0 followers case.
-    public int calculate(int followers, int publicRepos) {
-        return followers == 0 ? 0 : 6 / followers * publicRepos * 2;
+    public double calculate(int followers, int publicRepos) {
+        return (followers == 0 ? 0 : (double) 6 / followers) * publicRepos * 2;
     }
 }
